Added items for the force constants and need help getting the input files files for pytest.
